+++ code/renderervk_cplus/tr_model_iqm.cpp
@@ -24,6 +24,10 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 
 #include "tr_model_iqm.hpp"
 #include "tr_surface.hpp"
+#include "tr_main.hpp"
+#include "tr_shader.hpp"
+#include "tr_light.hpp"
+#include "tr_image.hpp"
 
 #define LL(x) x = LittleLong(x)
 
@@ -217,13 +221,13 @@ void R_AddIQMSurfaces(trRefEntity_t *ent)
 		// we will add shadows even if the main object isn't visible in the view
 
 		// stencil shadows can't do personal models unless I polyhedron clip
-		if (!personalModel && r_shadows->integer == 2 && fogNum == 0 && !(ent->e.renderfx & (RF_NOSHADOW | RF_DEPTHHACK)) && shader->sort == SS_OPAQUE)
+		if (!personalModel && r_shadows->integer == 2 && fogNum == 0 && !(ent->e.renderfx & (RF_NOSHADOW | RF_DEPTHHACK)) && shader->sort == static_cast<float>(SS_OPAQUE))
 		{
 			R_AddDrawSurf(reinterpret_cast<surfaceType_t *>(surface), tr.shadowShader, 0, 0);
 		}
 
 		// projection shadows work fine with personal models
-		if (r_shadows->integer == 3 && fogNum == 0 && (ent->e.renderfx & RF_SHADOW_PLANE) && shader->sort == SS_OPAQUE)
+		if (r_shadows->integer == 3 && fogNum == 0 && (ent->e.renderfx & RF_SHADOW_PLANE) && shader->sort == static_cast<float>(SS_OPAQUE))
 		{
 			R_AddDrawSurf(reinterpret_cast<surfaceType_t *>(surface), tr.projectionShadowShader, 0, 0);
 		}
@@ -420,8 +424,8 @@ static void ComputeJointMats(iqmData_t *data, int frame, int oldframe,
 }
 
 int R_IQMLerpTag(orientation_t *tag, iqmData_t *data,
-					  int startFrame, int endFrame,
-					  float frac, const char *tagName)
+				 int startFrame, int endFrame,
+				 float frac, const char *tagName)
 {
 	float jointMats[IQM_MAX_JOINTS * 12];
 	int joint;
